refactor: migrate to BigDecimal and JPA auto-update for development
- Convert all financial fields (Double -> BigDecimal) in entities, DTOs, services

- Add missing OrderItem.price field

- Disable Flyway for development flexibility

- Enable JPA ddl-auto=update mode

- Update V1 migration schema (Orders, Shipments, OrderStatusHistory)

- Fix JWT validation and authentication filter

- Clean up unused imports and refactor service dependencies

+++ backend/src/main/java/iuh/fit/backend/dto/CartItemResponse.java
@@ -1,5 +1,7 @@
 package iuh.fit.backend.dto;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -13,6 +15,6 @@ public class CartItemResponse {
     private String productName;
     private String variantName;
     private Integer quantity;
-    private Double price;
-    private Double subtotal;
+    private BigDecimal price;
+    private BigDecimal subtotal;
 }
