+++ backend/src/main/java/iuh/fit/backend/dto/CouponResponse.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -15,8 +16,8 @@ public class CouponResponse {
     private String description;
     private Boolean isActive;
     private String discountType;
-    private Double discountValue;
-    private Double minOrderValue;
+    private BigDecimal discountValue;
+    private BigDecimal minOrderValue;
     private Integer maxUsageValue;
     private LocalDateTime validFrom;
     private LocalDateTime validTo;
