+++ backend/src/main/java/iuh/fit/backend/dto/OrderResponse.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -14,10 +15,10 @@ public class OrderResponse {
     private Integer id;
     private Integer userId;
     private String status;
-    private Double subtotal;
-    private Double totalAmount;
-    private Double discountAmount;
-    private Double shippingFee;
+    private BigDecimal subtotal;
+    private BigDecimal totalAmount;
+    private BigDecimal discountAmount;
+    private BigDecimal shippingFee;
     private String notes;
     private LocalDate estimateDeliveryFrom;
     private LocalDate estimateDeliveryTo;
